+++ GRDB/Core/DatabaseQueue.swift
@@ -203,6 +203,8 @@ extension DatabaseQueue {
         return try writer.reentrantSync(block)
     }
     
+    /// This method is deprecated. Use concurrentRead instead.
+    ///
     /// Synchronously executes *block*.
     ///
     /// This method must be called from the protected database dispatch queue,
@@ -211,16 +213,30 @@ extension DatabaseQueue {
     /// Starting SQLite 3.8.0 (iOS 8.2+, OSX 10.10+, custom SQLite builds and
     /// SQLCipher), attempts to write in the database from this meethod throw a
     /// DatabaseError of resultCode `SQLITE_READONLY`.
-    ///
-    /// See `DatabaseWriter.readFromCurrentState`.
-    ///
-    /// :nodoc:
+    @available(*, deprecated, message: "Use concurrentRead instead")
     public func readFromCurrentState(_ block: @escaping (Database) -> Void) {
         // Check that we're on the correct queue...
         writer.execute { db in
             // ... and that no transaction is opened.
             GRDBPrecondition(!db.isInsideTransaction, "readFromCurrentState must not be called from inside a transaction.")
             db.readOnly { block(db) }
+        }
+    }
+    
+    public func concurrentRead<T>(_ block: @escaping (Database) throws -> T) -> Future<T> {
+        // DatabaseQueue can't perform parallel reads.
+        let result = Result<T> {
+            // Check that we're on the writer queue...
+            try writer.execute { db in
+                // ... and that no transaction is opened.
+                GRDBPrecondition(!db.isInsideTransaction, "concurrentRead must not be called from inside a transaction.")
+                return try db.readOnly {
+                    try block(db)
+                }
+            }
+        }
+        return Future {
+            try result.unwrap()
         }
     }