+++ GRDB/Core/DatabaseSnapshot.swift
@@ -34,7 +34,7 @@ public class DatabaseSnapshot : DatabaseReader {
             try db.beginTransaction(.deferred)
             
             // Take snapshot
-            // See DatabasePool.readFromCurrentState for a complete discussion
+            // See DatabasePool.concurrentRead for a complete discussion
             try db.makeSelectStatement("SELECT rootpage FROM sqlite_master").makeCursor().next()
         }
     }